emscripten update

+++ Client/Socket.c
@@ -37,7 +37,7 @@ static uint32_t at = 0;
 #ifdef EMSCRIPTEN
 void rr_on_socket_event_emscripten(struct rr_websocket *this,
                                    enum rr_websocket_event_type type,
-                                   void *data, uint64_t data_size)
+                                   void *data, uint32_t data_size)
 {
     rr_game_websocket_on_event_function(type, data, this->user_data, data_size);
 }
